+++ fist-kit-infra/fist-jasypt/fist-jasypt-core/src/main/java/com/power4j/fist/jasypt/core/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.power4j.fist.jasypt.core;
+
+import org.jspecify.annotations.NullMarked;