+++ fist-kit-infra/fist-jasypt/fist-jasypt-spring-boot-starter/src/main/java/com/power4j/fist/jasypt/boot/autoconfigure/package-info.java
@@ -1,0 +1,2 @@
+@org.jspecify.annotations.NullMarked
+package com.power4j.fist.jasypt.boot.autoconfigure;